+++ packages/devtools_app/assets/img/layout_explorer/cross_axis_alignment/column_stretch.png
