+++ packages/devtools_app/assets/img/layout_explorer/free_space.png
