+++ packages/devtools_app/assets/img/layout_explorer/main_axis_alignment/center.png
