+++ packages/devtools_app/lib/src/inspector/flutter/inspector_data_models.dart
@@ -10,7 +10,7 @@ import 'package:flutter/widgets.dart';
 import '../../utils.dart';
 import '../diagnostics_node.dart';
 import '../enum_utils.dart';
-import 'layout_explorer/utils.dart';
+import 'layout_explorer/flex/utils.dart';
 
 const overflowEpsilon = 0.1;
