+++ packages/devtools_app/lib/src/inspector/flutter/layout_explorer/flex/flex.dart
@@ -9,14 +9,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
-import '../../../ui/colors.dart';
-import '../../../ui/theme.dart';
-import '../../../utils.dart';
-import '../../diagnostics_node.dart';
-import '../../inspector_controller.dart';
-import '../../inspector_service.dart';
-import '../inspector_data_models.dart';
-import '../inspector_service_flutter_extension.dart';
+import '../../../../ui/colors.dart';
+import '../../../../ui/theme.dart';
+import '../../../../utils.dart';
+import '../../../diagnostics_node.dart';
+import '../../../inspector_controller.dart';
+import '../../../inspector_service.dart';
+import '../../inspector_data_models.dart';
+import '../../inspector_service_flutter_extension.dart';
 import 'arrow.dart';
 import 'free_space.dart';
 import 'overflow_indicator_painter.dart';
