+++ packages/devtools_app/lib/src/inspector/flutter/layout_explorer/flex/overflow_indicator_painter.dart
