+++ packages/devtools_app/test/flutter/inspector_screen_test.dart
@@ -7,8 +7,8 @@ import 'dart:convert';
 import 'package:devtools_app/src/globals.dart';
 import 'package:devtools_app/src/inspector/diagnostics_node.dart';
 import 'package:devtools_app/src/inspector/flutter/inspector_screen.dart';
-import 'package:devtools_app/src/inspector/flutter/inspector_screen_details_tab.dart';
-import 'package:devtools_app/src/inspector/flutter/layout_explorer/flex.dart';
+import 'package:devtools_app/src/inspector/flutter/layout_explorer/flex/flex.dart';
+import 'package:devtools_app/src/inspector/flutter/layout_explorer/layout_explorer.dart';
 import 'package:devtools_app/src/inspector/inspector_controller.dart';
 import 'package:devtools_app/src/inspector/inspector_service.dart';
 import 'package:devtools_app/src/inspector/inspector_tree.dart';
