+++ packages/devtools_app/test/flutter/layout_explorer/flex/arrow_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:devtools_app/src/inspector/flutter/layout_explorer/arrow.dart';
+import 'package:devtools_app/src/inspector/flutter/layout_explorer/flex/arrow.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
