+++ packages/devtools_app/test/flutter/layout_explorer/flex/goldens/arrow_bidirectional_horizontal.png
