+++ packages/devtools_app/test/flutter/layout_explorer/flex/goldens/arrow_unidirectional_left.png
