+++ packages/devtools_app/test/flutter/layout_explorer/flex/goldens/story_of_column_layout.png
